Benutzer ermitteln mit Prüfung auf gleichen Träger des Users

+++ wbs-alarm/wbs-service/src/main/java/de/hsfulda/et/wbs/repository/BenutzerRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.math.BigInteger;
 
-public interface BenutzerRepository extends CrudRepository<Benutzer, BigInteger> {
+public interface BenutzerRepository extends CrudRepository<Benutzer, Long> {
 
     Benutzer findByUsername(String username);
 }
